requirements.txt
bitmath
iso-639
PyYAML
